refactored CNMFParams into its own file. Changed the way params are accessed and set.
